requirements.txt
fastapi
uvicorn
PyPDF2
langchain
langchain-community
python-dotenv
pyttsx3
SpeechRecognition
faiss-cpu
tiktoken
openai
sentence-transformers 
transformers 
python-multipart